Move test into generated test files in test directories
