+++ test/controllers/events_controller_test.rb
@@ -1,7 +1,9 @@
-require 'test_helper'
+require_relative '../test_helper'
 
 class EventsControllerTest < ActionController::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
+  test "should get index" do
+    get :index
+    assert_response :success
+    assert_not_nil assigns(:events)
+  end
 end
