--- test/unit/events_controller_test.rb
@@ -1,9 +1,0 @@
-require '../test_helper'
-
-class EventsControllerTest < ActionController::TestCase
-  test "should get index" do
-    get :index
-    assert_response :success
-    assert_not_nil assigns(:events)
-  end
-end
